formalize new opto tasks
