+++ ibllib/io/extractors/extractor_types.json
@@ -1,15 +1,15 @@
 {
+ "opto_ephysChoiceWorld": "ephys_biased_opto",
  "ephysChoiceWorld": "ephys",
  "ephyskarolinaChoiceWorld": "ephys",
  "trainingChoiceWorldWidefield": "ephys",
  "passive_opto": "ephys",
- "opto_ephysChoiceWorld": "ephys_biased",
  "widefieldChoiceWorld": "ephys",
+ "opto_biasedChoiceWorld": "biased_opto",
  "_biasedChoiceWorld": "biased",
  "biasedScanningChoiceWorld": "biased",
  "biasedVisOffChoiceWorld": "biased",
  "karolinaChoiceWorld": "biased",
- "opto_biasedChoiceWorld": "biased_opto",
  "_habituationChoiceWorld": "habituation",
  "_trainingChoiceWorld": "training",
  "ephysMockChoiceWorld": "mock_ephys",
