+++ ibllib/io/extractors/training_trials.py
@@ -647,23 +647,6 @@ class StimOnOffFreezeTimes(BaseBpodTrialsExtractor):
         return stimOn_times, stimOff_times, stimFreeze_times
 
 
-class LaserBool(BaseBpodTrialsExtractor):
-    save_names = ['_ibl_trials.laser_stimulation.npy', '_ibl_trials.laser_probability.npy']
-    var_names = ['laser_stimulation', 'laser_probability']
-
-    def _extract(self):
-        lstim = np.array([np.float(t.get('laser_stimulation', np.NaN)) for t in self.bpod_trials])
-        lprob = np.array([np.float(t.get('laser_probability', np.NaN)) for t in self.bpod_trials])
-        _logger.info('Extracting laser datasets')
-        if np.all(np.isnan(lprob)):
-            self.save_names[1] = None  # this prevents the file from being saved when no data
-            _logger.info('No laser probability found in bpod data')
-        if np.all(np.isnan(lstim)):
-            self.save_names[0] = None  # this prevents the file from being saved when no data
-            _logger.info('No laser stimulation found in bpod data')
-        return lstim, lprob
-
-
 def extract_all(session_path, save=False, bpod_trials=None, settings=None):
     if not bpod_trials:
         bpod_trials = raw.load_data(session_path)
@@ -672,7 +655,7 @@ def extract_all(session_path, save=False, bpod_trials=None, settings=None):
     if settings is None or settings['IBLRIG_VERSION_TAG'] == '':
         settings = {'IBLRIG_VERSION_TAG': '100.0.0'}
 
-    base = [FeedbackType, ContrastLR, ProbabilityLeft, Choice, RepNum, RewardVolume, LaserBool,
+    base = [FeedbackType, ContrastLR, ProbabilityLeft, Choice, RepNum, RewardVolume,
             FeedbackTimes, Intervals, ResponseTimes, GoCueTriggerTimes, GoCueTimes]
     # Version check
     if version.ge(settings['IBLRIG_VERSION_TAG'], '5.0.0'):
